% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{coverage}
\alias{coverage}
\title{Measure the coverage probability}
\usage{
coverage(ci)
}
\arguments{
\item{ci}{The result of CI calculation.}
}
\description{
This function is to measure the coverage probability.
}
