% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_sequentially_max_likelihood}
\alias{estimate_sequentially_max_likelihood}
\title{Parameter estimation by sequential execution to get max loglikelihood}
\usage{
estimate_sequentially_max_likelihood(
  dataset,
  verbose = TRUE,
  maximize_num = 100L,
  theta_Ic_init = rep(0.65, 100L),
  theta_It_init = rep(0.75, 100L),
  lambda_Ic_nr_init = rep(0.1, 100L),
  lambda_It_nr_init = rep(0.05, 100L),
  lambda_IcMc_r_init = rep(0.1, 100L),
  lambda_IcMc_nr_init = rep(0.2, 100L),
  lambda_ItMc_r_init = rep(0.08, 100L),
  lambda_ItMc_nr_init = rep(0.16, 100L),
  lambda_IcMt_r_init = rep(0.08, 100L),
  lambda_IcMt_nr_init = rep(0.06, 100L),
  lambda_ItMt_r_init = rep(0.04, 100L),
  lambda_ItMt_nr_init = rep(0.08, 100L),
  max_iter = 5000L,
  eps = 1e-05,
  option = NULL
)
}
\arguments{
\item{dataset}{The datasets of scenarios from function generate_scenario.}

\item{verbose}{If TRUE, show progress.}

\item{maximize_num}{The number of iteration.}

\item{theta_Ic_init}{Initial value vectors of proportion of responders in induction control.}

\item{theta_It_init}{Initial value vectors of proportion of responders in induction treatment.}

\item{lambda_Ic_nr_init}{Initial value vectors of hazard of non-responders in induction control.}

\item{lambda_It_nr_init}{Initial value vectors of hazard of non-responders in induction treatment.}

\item{lambda_IcMc_r_init}{Initial value vectors of hazard of responders in maintenance control when induction control.}

\item{lambda_IcMc_nr_init}{Initial value vectors of hazard of non-responders in maintenance control when induction control.}

\item{lambda_ItMc_r_init}{Initial value vectors of hazard of responders in maintenance treatment when induction control.}

\item{lambda_ItMc_nr_init}{Initial value vectors of hazard of non-responders in maintenance treatment when induction control.}

\item{lambda_IcMt_r_init}{Initial value vectors of hazard of responders in maintenance control when induction treatment.}

\item{lambda_IcMt_nr_init}{Initial value vectors of hazard of non-responders in maintenance control when induction treatment.}

\item{lambda_ItMt_r_init}{Initial value vectors of hazard of responders in maintenance treatment when induction treatment.}

\item{lambda_ItMt_nr_init}{Initial value vectors of hazard of non-responders in maintenance treatment when induction treatment.}

\item{max_iter}{The maximum number of iteration in each EM algorithm.}

\item{eps}{The threshold of difference log-likelihoods to stop EM algorithm.}

\item{option}{If NULL, implement default EM algorithm.
If "fix_theta", implement EM algorithm being theta fixed.}
}
\description{
This function is to estimate parameters by sequential execution.
Full dataset from generate_scenario is used in estimation.
}
