% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-dataset.R
\name{generate_scenario}
\alias{generate_scenario}
\title{Generate datasets in one scenario}
\usage{
generate_scenario(
  sim_num = 1000L,
  t_judge = 6,
  censor_rate = 0,
  IcMc_size = 1000L,
  ItMc_size = 1000L,
  IcMt_size = 1000L,
  ItMt_size = 1000L,
  theta_Ic = 0.65,
  theta_It = 0.75,
  lambda_Ic_r = 0,
  lambda_Ic_nr = 0.1,
  lambda_It_r = 0,
  lambda_It_nr = 0.05,
  lambda_IcMc_r = 0.1,
  lambda_IcMc_nr = 0.2,
  lambda_ItMc_r = 0.08,
  lambda_ItMc_nr = 0.16,
  lambda_IcMt_r = 0.08,
  lambda_IcMt_nr = 0.16,
  lambda_ItMt_r = 0.04,
  lambda_ItMt_nr = 0.08,
  seed = 42
)
}
\description{
This function is to generate datasets with true parameters in each treatment group.
Each dataset consists of survival time, censoring status, and treatment assignment.
}
