% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_tsrd_dataset}
\alias{sample_tsrd_dataset}
\title{Sampling times from the survival function}
\usage{
sample_tsrd_dataset(
  n,
  t_judge,
  theta,
  lambda_I_r,
  lambda_I_nr,
  lambda_IM_r,
  lambda_IM_nr
)
}
\description{
This function is to sample from survival function in specified parameters.
The sampling method is inverse transform sampling.
}
