% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-dataset.R
\name{generate_scenario}
\alias{generate_scenario}
\title{Generate datasets in one scenario}
\usage{
generate_scenario(
  sim_num = 1000L,
  t_judge = 6,
  censor_rate = 0,
  IcMc_size = 1000L,
  ItMc_size = 1000L,
  IcMt_size = 1000L,
  ItMt_size = 1000L,
  theta_Ic = 0.65,
  theta_It = 0.75,
  lambda_Ic_r = 0,
  lambda_Ic_nr = 0.1,
  lambda_It_r = 0,
  lambda_It_nr = 0.05,
  lambda_IcMc_r = 0.1,
  lambda_IcMc_nr = 0.2,
  lambda_ItMc_r = 0.08,
  lambda_ItMc_nr = 0.16,
  lambda_IcMt_r = 0.08,
  lambda_IcMt_nr = 0.16,
  lambda_ItMt_r = 0.04,
  lambda_ItMt_nr = 0.08,
  seed = 42
)
}
\arguments{
\item{sim_num}{The number of simulations.}

\item{t_judge}{The time to judge whether to be responder or not.}

\item{censor_rate}{The proportion of censoring.}

\item{IcMc_size}{The sample size of induction control and maintenance control.}

\item{ItMc_size}{The sample size of induction treatment and maintenance control.}

\item{IcMt_size}{The sample size of induction control and maintenance treatment.}

\item{ItMt_size}{The sample size of induction treatment and maintenance treatment.}

\item{theta_Ic}{The proportion of responders in induction control.}

\item{theta_It}{The proportion of responders in induction treatment.}

\item{lambda_Ic_r}{The hazard of responders in induction control. This should be set to 0.}

\item{lambda_Ic_nr}{The hazard of non-responders in induction control.}

\item{lambda_It_r}{The hazard of responders in induction treatment. This should be set to 0.}

\item{lambda_It_nr}{The hazard of non-responders in induction treatment.}

\item{lambda_IcMc_r}{The hazard of responders in maintenance control when induction control.}

\item{lambda_IcMc_nr}{The hazard of non-responders in maintenance control when induction control.}

\item{lambda_ItMc_r}{The hazard of responders in maintenance treatment when induction control.}

\item{lambda_ItMc_nr}{The hazard of non-responders in maintenance treatment when induction control.}

\item{lambda_IcMt_r}{The hazard of responders in maintenance control when induction treatment.}

\item{lambda_IcMt_nr}{The hazard of non-responders in maintenance control when induction treatment.}

\item{lambda_ItMt_r}{The hazard of responders in maintenance treatment when induction treatment.}

\item{lambda_ItMt_nr}{The hazard of non-responders in maintenance treatment when induction treatment.}

\item{seed}{The value of random seed.}
}
\description{
This function is to generate datasets with true parameters in each treatment group.
Each dataset consists of survival time, censoring status, and treatment assignment.
}
