% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_tsrd_dataset}
\alias{sample_tsrd_dataset}
\title{Sampling times from the survival function}
\usage{
sample_tsrd_dataset(
  n,
  t_judge,
  theta,
  lambda_I_r,
  lambda_I_nr,
  lambda_IM_r,
  lambda_IM_nr
)
}
\arguments{
\item{n}{The number of sample.}

\item{t_judge}{The time to judge whether to be responder or not.}

\item{theta}{The proportion of responders.}

\item{lambda_I_r}{The hazard of responders in induction.}

\item{lambda_I_nr}{The hazard of non-responders in induction.}

\item{lambda_IM_r}{The hazard of responders in maintenance.}

\item{lambda_IM_nr}{The hazard of non-responders in maintenance.}
}
\description{
This function is to sample from survival function in specified parameters.
The sampling method is inverse transform sampling.
}
